fix: rename module to hnn_io because io conflicts with python built-in module and causes odd debugging behavior.
