+++ hnn_core/__init__.py
@@ -6,6 +6,6 @@ from .cell import Cell
 from .cell_response import CellResponse, read_spikes
 from .cells_default import pyramidal, basket
 from .parallel_backends import MPIBackend, JoblibBackend
-from .io import read_network
+from .hnn_io import read_network
 
 __version__ = '0.4.dev0'
